[ch88739] Fix decoding of multiple events in a single chunk (#6)
Rewrite the SSE decoder again to fix a nasty bug that seemed to be occurring intermittently during IH tests (and in any case was reproducible). If two events occurred within a single chunk, we would drop one, which is unfortunate. This may have been introduced during the last rewrite (#3).

I also tried hard to simplify down the logic and remove some conditionals that didn't need to be there, so hopefully the result is a little more readable than what came before.