Fix usage of split_inclusive in event parser logic
It was relying on old behavior where Some(empty slice) would be returned
if there were no matches. This was updated in more recent rust versions to return None,
but CircleCI was using a much older version of rust than I have locally.

Updated the usage and CircleCI config to use cimg/rust.