Do not initialize event data on comments (#38)
As we process a stream, we build up the event_data instance with the
data as it is parsed. When we receive an empty line (e.g. "\n\r"), we
try to dispatch any event we have pending.

However, this event_data instance should only be created and / or
updated if the line we are parsing isn't a comment. Otherwise, we can
skip the initialization and return a comment event without furthering
processing.