ci: Separate CI steps
The contract tests require a secondary set of dependencies in order to
spin up the required webserver. As a result, the required minimum rustc
version might change independently of changes to our library.

Similarly, generating a code coverage report requires a unique set of
crates as well.

To help with this, building the core library, generating coverage
reports, and running the contract tests have been moved into separate CI
jobs.